add: watcher link logic,new slider components to pages fix: preloader full screen problem

+++ resources/js/router/middlewares/AuthCheckProfileMiddleware.js
@@ -1,12 +1,12 @@
 import store from "@/store";
 
-export async function AuthCheckProfileMiddleware(_, router) {
+export async function AuthCheckProfileMiddleware(route, router) {
     const user = store.getters.user;
 
-    if (!!user && Object.entries(user).length === 0) {
+    if (!!user && Object.entries(user).length === 0 && !route?.query?.access_key) {
         router.push({ name: "home" });
         store.dispatch("drop_all");
     } else {
-        store.dispatch("set_accounts", user.id);
+        store.dispatch("set_accounts", { route: route, user_id: user.id });
     }
 }
